created interfaces for cache and parser
